backend/pyproject.toml
[project]
name = "vibeshift-backend"
version = "0.1.0"
description = "VibeShift backend service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  # 为了 Docker 构建更稳定更快，这里锁定到已验证可用的版本，避免 pip 解析器长时间回溯。
  "fastapi==0.135.1",
  "uvicorn[standard]==0.42.0",
  "pydantic==2.12.5",
  "pydantic-settings==2.13.1",
  "httpx==0.28.1",
  "trafilatura==2.0.0",
  "readability-lxml==0.8.4.1",
  "beautifulsoup4==4.14.3",
  "openai==1.109.1",
  "duckduckgo-search==7.5.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
